requirements.txt
amqp==5.2.0
appdirs==1.4.4
arabic-reshaper==3.0.0
asgi-redis==1.4.3
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
baron==0.10.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.3
cssselect2==0.7.0
daphne==4.1.2
defusedxml==0.8.0rc2
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==5.1
django-appconf==1.0.6
django-autoslug==1.9.9
django-cache-url==3.4.5
django-celery-email==3.0.0
django-cors-headers==4.4.0
django-countries==7.6.1
django-environ==0.11.2
django-environs==0.1.1
django-filter==24.3
django-phonenumber-field==8.0.0
django-templated-mail==1.1.1
django-user-sessions==2.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
environs==11.0.0
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.2
lxml==5.3.0
marshmallow==3.21.3
msgpack==1.1.0
msgpack-python==0.5.6
oauthlib==3.2.2
oscrypto==1.3.0
packaging==24.1
phonenumbers==8.13.42
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.9.0
pyOpenSSL==24.2.1
pypdf==5.1.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
qrcode==8.0
redbaron==0.9.2
redis==5.2.0
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rply==0.7.8
service-identity==24.2.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.4.0
tomli==2.1.0
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
uvicorn==0.32.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.8.2
xhtml2pdf==0.2.16
zope.interface==7.1.1
